Refactor credentials check
